fix: reboot HVAC ESPs after prolonged Wi-Fi loss
Add a 5-minute STA offline watchdog so devices recover when the fallback AP leaves them stranded, and cast dump_config interval logs to silence ESP32-C6 format warnings.

+++ esphome/components/midea_xye/climate_midea_xye.cpp
@@ -763,8 +763,8 @@ climate::ClimateTraits ClimateMideaXYE::traits() {
 
 void ClimateMideaXYE::dump_config() {
   ESP_LOGCONFIG(Constants::TAG, "MideaXYE:");
-  ESP_LOGCONFIG(Constants::TAG, "  [x] Period: %dms", this->get_update_interval());
-  ESP_LOGCONFIG(Constants::TAG, "  [x] Response timeout: %dms", this->response_timeout);
+  ESP_LOGCONFIG(Constants::TAG, "  [x] Period: %dms", (int) this->get_update_interval());
+  ESP_LOGCONFIG(Constants::TAG, "  [x] Response timeout: %dms", (int) this->response_timeout);
   ESP_LOGCONFIG(Constants::TAG, "  [x] Use Fahrenheit: %d", this->use_fahrenheit_);
 
 #ifdef USE_REMOTE_TRANSMITTER
